AlbumListView: Allow subclasses to switch between album grid and list
Add SwitchToAlbumList and SwitchToAlbumGrid protected methods, so that
AlbumListView subclasses can choose whether to display the album grid
or the old album list.

The "Disable album grid" preference is not changed by these methods, and
it still determines the initial display. This means subclasses must be
careful if they don't want to follow the preference.