InternetRadio: Previous should never mean a restart (bgo#665268)
In the context of a radio station, as it is not seekable, the button
"PreviousOrRestart" should always mean Previous.

Furthermore, simple testing seemed to suggest that Position was never
returning the duration that the radio had already been playing
(probably because it's a non seekable resource), so we twist the
logic upside down to make it a bit more readable.

Signed-off-by: Andres G. Aragoneses <knocte@gmail.com>